Allow starting the scala server with test:reStart
