+++ interop-tests/src/test/scala/com/lightbend/grpc/interop/app/AkkaHttpServerAppScala.scala
@@ -1,0 +1,7 @@
+package com.lightbend.grpc.interop.app
+
+import com.lightbend.grpc.interop.AkkaHttpServerProviderScala
+
+object AkkaHttpServerAppScala extends App {
+  AkkaHttpServerProviderScala.server.start()
+}
