+++ project/plugins.sbt
@@ -8,6 +8,7 @@ addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.14.6")
 addSbtPlugin("com.eed3si9n" % "sbt-buildinfo" % "0.8.0")
 addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.9.0")
 addSbtPlugin("com.lightbend.akka" % "sbt-paradox-akka" % "0.6")
+addSbtPlugin("io.spray" % "sbt-revolver" % "0.9.1")
 
 // scripted testing
 libraryDependencies += "org.scala-sbt" %% "scripted-plugin" % sbtVersion.value
